// frontend/src/generated/types/paymentCoupon.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * QROrder API
 * QROrder Management System API 문서
 * OpenAPI spec version: v1.0
 */

export interface PaymentCoupon {
  sysId?: string;
  couponCd?: string;
  couponNm?: string;
  startDate?: string;
  endDate?: string;
  useYn?: string;
}